Comments, and initial work on interaction token tests
