+++ ts/credentials/signedCredential/signedCredential.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata'
 import { Transform, plainToClass, classToPlain, Type, Exclude, Expose } from 'class-transformer'
 import { canonize } from 'jsonld'
 import { generateRandomID, sha256 } from '../../utils/crypto'
-import { ISignedCredentialAttrs } from './types'
+import { ISignedCredentialAttrs, ISignedCredCreationArgs } from './types'
 import { ILinkedDataSignature, IDigestable } from '../../linkedDataSignature/types'
 import { ContextEntry, BaseMetadata } from 'cred-types-jolocom-core'
 import { IClaimSection } from '../credential/types'
