+++ ts/identityWallet/types.ts
@@ -1,6 +1,6 @@
 import { Identity } from '../identity/identity'
-import { IVaultedKeyProvider } from '../crypto/softwareProvider'
-import { IKeyMetadata } from './identityWallet'
+import { IVaultedKeyProvider } from '../vaultedKeyProvider/softwareProvider'
+import { IKeyMetadata } from '../credentials/signedCredential/types'
 
 export interface IIdentityWalletCreateArgs {
   vaultedKeyProvider: IVaultedKeyProvider
