--- ts/interactionFlows/authentication/types.ts
@@ -1,17 +1,0 @@
-export interface IAuthenticationAttrs {
-  challenge: string
-  callbackURL: string
-}
-
-export interface IAuthPayloadCreationAttrs {
-  iss?: string
-  typ: string
-  authentication: IAuthenticationAttrs
-}
-
-export interface IAuthentiactionPayloadAttrs {
-  iat: number
-  iss: string
-  typ: string
-  authentication: IAuthenticationAttrs
-}
