+++ ts/interactionFlows/credentialResponse.ts
@@ -1,7 +1,7 @@
 import { plainToClass, classToPlain } from 'class-transformer'
 import { ICredentialResponseAttrs } from './credentialResponse/types'
 import { ISignedCredentialAttrs } from '../credentials/signedCredential/types'
-import { SignedCredential } from '../credentials/signedCredential/signedCredential'
+import { SignedCredential } from '../credentials/signedCredential/signedCredential'
 import { CredentialRequest } from './credentialRequest'
 
 export class CredentialResponse {
@@ -9,8 +9,7 @@ export class CredentialResponse {
 
   public static create(credentials: ISignedCredentialAttrs[]): CredentialResponse {
     const credentialResponse = new CredentialResponse()
-    credentialResponse.suppliedCredentials = credentials
-      .map((sCred) => plainToClass(SignedCredential, sCred))
+    credentialResponse.suppliedCredentials = credentials.map(sCred => plainToClass(SignedCredential, sCred))
 
     return credentialResponse
   }
@@ -20,8 +19,7 @@ export class CredentialResponse {
   }
 
   public satisfiesRequest(cr: CredentialRequest): boolean {
-    const credentials = this.suppliedCredentials
-      .map((sCredClass) => sCredClass.toJSON())
+    const credentials = this.suppliedCredentials.map(sCredClass => sCredClass.toJSON())
 
     const validCredentials = cr.applyConstraints(credentials)
     return this.suppliedCredentials.length === validCredentials.length
@@ -33,8 +31,7 @@ export class CredentialResponse {
 
   public static fromJSON(json: ICredentialResponseAttrs): CredentialResponse {
     const credResponse = plainToClass(CredentialResponse, json)
-    credResponse.suppliedCredentials = json.suppliedCredentials
-      .map((sCred) => plainToClass(SignedCredential, sCred))
+    credResponse.suppliedCredentials = json.suppliedCredentials.map(sCred => plainToClass(SignedCredential, sCred))
 
     return credResponse
   }
